# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RBAC scaffolding based on Gin + Gorm + Casbin + Dig.",
        "title": "cappuccino",
        "contact": {},
        "license": {},
        "version": "5.2.1"
    },
    "host": "127.0.0.1:1207",
    "basePath": "/",
    "paths": {
        "/api/v1/demos": {
            "get": {
                "tags": [
                    "Demo"
                ],
                "summary": "查询数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "分页索引",
                        "name": "current",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "编号",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态(1:启用 2:停用)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询结果：{list:列表数据,pagination:{current:页索引,pageSize:页大小,total:总数量}}",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.Demo"
                            }
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Demo"
                ],
                "summary": "创建数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "创建数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Demo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Demo"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/demos/{id}": {
            "get": {
                "tags": [
                    "Demo"
                ],
                "summary": "查询指定数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Demo"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "404": {
                        "description": "{error:{code:0,message:资源不存在}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Demo"
                ],
                "summary": "更新数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Demo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Demo"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Demo"
                ],
                "summary": "删除数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/demos/{id}/disable": {
            "patch": {
                "tags": [
                    "Demo"
                ],
                "summary": "禁用数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/demos/{id}/enable": {
            "patch": {
                "tags": [
                    "Demo"
                ],
                "summary": "启用数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "schema.Demo": {
            "type": "object",
            "required": [
                "code",
                "name",
                "status"
            ],
            "properties": {
                "code": {
                    "description": "编号",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "record_id": {
                    "description": "记录ID",
                    "type": "string"
                },
                "status": {
                    "description": "状态(1:启用 2:停用)",
                    "type": "integer"
                }
            }
        },
        "schema.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误项",
                    "type": "object",
                    "$ref": "#/definitions/schema.HTTPErrorItem"
                }
            }
        },
        "schema.HTTPErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "schema.HTTPStatus": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "状态(OK)",
                    "type": "string"
                }
            }
        },
...
    }
}